% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{add_centroid}
\alias{add_centroid}
\title{\code{get_centroid}}
\usage{
add_centroid(df)
}
\arguments{
\item{df}{an object of class \code{sf}}
}
\description{
Geometric operation on simple feature geometry to add coordinates for feature centroid.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
