% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_matrices.R
\name{add_mmatrix}
\alias{add_mmatrix}
\title{Add Mismatching Matrices}
\usage{
add_mmatrix(x, y)
}
\arguments{
\item{x}{a matrix-like object}

\item{y}{a matrix-like object}
}
\description{
Add matrices of different sizes.
}
