% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nsd.R
\name{calc_nsd}
\alias{calc_nsd}
\title{Calculate Net Squared Displacement}
\usage{
calc_nsd(x, y, normalize = FALSE)
}
\arguments{
\item{x}{A vector of the UTM Easting coordinate}

\item{y}{A vector of the UTM Northing coordinate}

\item{normalize}{A logical value that indicates whether to normalize the result. The default is FALSE}
}
\value{
A vector of the NSD values for each (x, y) vector pair.
}
\description{
Calculate Net Squared Displacement
}
