% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perimeter.R
\name{calc_perimeter}
\alias{calc_perimeter}
\title{Calculate the perimeter of a polygon}
\usage{
calc_perimeter(shape)
}
\arguments{
\item{shape}{The XY coordinates of a polygon perimeter}
}
\value{
The perimeter of the given polygon
}
\description{
\code{calc_perimeter} returns the perimeter of simple polygons. This function
is used internally to calculate the perimeter of the local convex hulls.
}
