% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print_header}
\alias{.print_header}
\title{Print header line for \code{proto_graph} object}
\usage{
.print_header(x, ...)
}
\arguments{
\item{x}{A class \code{proto_graph} object}

\item{...}{Extra arguments to pass to \code{print()}.}
}
\description{
Print header line for \code{proto_graph} object
}
\keyword{internal}
