% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_mbe.R
\name{gen_mbe}
\alias{gen_mbe}
\title{Generates a minimum bounding ellipse for a point set}
\usage{
gen_mbe(x, y)
}
\arguments{
\item{x}{A vector of the UTM Easting coordinates of a point set}

\item{y}{A vector of the UTM Northing coordinates of a point set}
}
\value{
An ellipse object (see \code{?ellipse}) with the added values of the
  eccentricity of the ellippse, and the semi-major and semi-minor axes.
}
\description{
\code{gen_mbe} generates a minimum bounding ellipse for the given point set.
A minimum bounding ellipse is the smallest possible ellipse that contains the entire point set.
}
