% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{get_geo_edges}
\alias{get_geo_edges}
\title{\code{get_geo_edges}}
\usage{
get_geo_edges(edges, nodes, .centroid_latitude, .centroid_longitude)
}
\arguments{
\item{edges}{...}

\item{nodes}{...}

\item{.centroid_latitude}{...}

\item{.centroid_longitude}{...}
}
\description{
Geometric operation on simple feature geometry to add coordinates for start and end point coordinates for each edge.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
