% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{get_proto_graph}
\alias{get_proto_graph}
\title{\code{get_proto_graph}}
\usage{
get_proto_graph(edges, nodes, directed = FALSE, centroid = TRUE)
}
\arguments{
\item{edges}{A \code{data.frame} containing a symbolic edge list in the first two columns. Additional columns are considered as edge attributes.}

\item{nodes}{A \code{data.frame} and object of class \code{sf} containing the node metadata. See details below.}

\item{directed}{Logical, whether or not to create a directed graph.}

\item{centroid}{Logical, whether or not to calculate the centroid for each geometry in the \code{nodes}. See details below.}
}
\description{
Produce a \code{proto_graph} object that can be laid as a network, map, or geocoded network on a map.
}
\details{
\code{nodes} must be a \code{data.frame} with vertex metadata. THe first column of the table is assume to contain the symbolic vertex names, this will be added to the graphs as the 'name' vertex attribute. Additionally, as an object of class \code{sf} it must include the intended geometry for each vertex.

\code{centroid} is used to calculate the geographic position of each node when the network is represented on a map. As a default, the argument is set to \code{TRUE} and will calculate the centroid for the provided geometry. If \code{FALSE} the user will have to provide a node latitude and longitude for graphing.
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
