% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_matrices.R
\name{intersect_mmatrix}
\alias{intersect_mmatrix}
\title{Intersect Mismatching Matrices}
\usage{
intersect_mmatrix(x, y)
}
\arguments{
\item{x}{a matrix-like object}

\item{y}{a matrix-like object}
}
\description{
Intersect matrices of different sizes.
}
