% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locoh.R
\name{locoh}
\alias{locoh}
\title{Creates a set of Local Convex Hulls}
\usage{
locoh(dat, cols = c("x", "y", "timestamp", "ndowid"))
}
\arguments{
\item{dat}{A \code{data.frame} that contains the XY, ID, and timestamp of
the dataset}

\item{cols}{A vector of column names for the following fields in the \code{data.frame}
x, y, timestamp, id. The default is \code{c('x', 'y', 'timestamp', 'id')}.
The entire vector must always be given if any default value is changed.}
}
\value{
A list of two \code{data.frames}. The first is a \code{data.frame} that
  contains the boundaries for all the LoCoHs. The second \code{data.frame}
  contains summary details for each LoCoH. These two \code{data.frames} share
  an id field, \code{id}.
}
\description{
\code{locoh} creates a set of Local Convex Hulls (LoCoH). LoCoHs are
minimum are MCPs around a selected point set from the dataset. The number
of points to include in a LoCoH defaults to 25 and can be controlled by
the \code{n_points} parameter (in dev). You will be able to select the
method by which LoCoH are created (k, r, a). \code{locoh} also generates a
minimum bounding ellipse for each LoCoH The summary details about each LoCoH
are added to the summary data.frame.
}
