% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{plot.proto_graph}
\alias{plot.proto_graph}
\title{proto_graph Print Method}
\usage{
\method{plot}{proto_graph}(x, ...)
}
\arguments{
\item{x}{A class \code{proto_graph} object}

\item{...}{Extra arguments to pass to \code{print()}.}
}
\description{
proto_graph Print Method
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
