% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.proto_graph}
\alias{print.proto_graph}
\title{\code{proto_graph} Print Method}
\usage{
\method{print}{proto_graph}(x, ...)
}
\arguments{
\item{x}{A class \code{proto_graph} object}

\item{...}{Extra arguments to pass to \code{print()}.}
}
\description{
\code{proto_graph} Print Method
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
