% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_matrices.R
\name{sub_mmatrix}
\alias{sub_mmatrix}
\title{Subsetting Mismatching Matrices}
\usage{
sub_mmatrix(subsetting_matrix, subsettor_matrix)
}
\arguments{
\item{subsetting_matrix}{a matrix-like object}

\item{subsettor_matrix}{a matrix-like object}
}
\description{
Subsetting Mismatching Matrices
}
