% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.proto_graph}
\alias{plot.proto_graph}
\title{\code{proto_graph} Plot Method}
\usage{
\method{plot}{proto_graph}(x, .xlab = "", .ylab = "", ...)
}
\arguments{
\item{x}{A class \code{proto_graph} object}

\item{...}{Extra arguments to pass to \code{print()}.}
}
\description{
\code{proto_graph} Plot Method
}
\author{
Christopher Callaghan, \email{cjcallag@nps.edu}
}
